+++ src/main/kotlin/jp/nephy/penicillin/Session.kt
@@ -3,22 +3,23 @@ package jp.nephy.penicillin
 import jp.nephy.jsonkt.JsonModel
 import jp.nephy.penicillin.model.Cursor
 import jp.nephy.penicillin.request.*
-import jp.nephy.penicillin.streaming.StreamHandler
+import jp.nephy.penicillin.request.streaming.StreamListener
 import okhttp3.OkHttpClient
 import java.util.concurrent.ScheduledThreadPoolExecutor
 import java.util.concurrent.TimeUnit
 
 
-class Session(val consumerKey: String?, val consumerSecret: String?, val accessToken: String?, val accessTokenSecret: String?, val isOfficialClient: Boolean, val bearerToken: String?, httpClientBuilder: OkHttpClient.Builder.() -> Unit, val maxRetries: Int) {
+class Session(val consumerKey: String?, val consumerSecret: String?, val accessToken: String?, val accessTokenSecret: String?, val isOfficialClient: Boolean, val bearerToken: String?, val knownDeviceToken: String?, httpClientBuilder: OkHttpClient.Builder.() -> Unit, optionBuilder: PenicillinOption.Builder.() -> Unit) {
     init {
         if (consumerKey == null && consumerSecret == null && accessToken == null && accessTokenSecret == null && bearerToken == null) {
             throw PenicillinLocalizedException(LocalizedString.CredentialsAreAllNull)
         }
     }
 
-    val httpClient = OkHttpClient.Builder().apply(httpClientBuilder).apply {
+    val httpClient = OkHttpClient.Builder().apply {
         readTimeout(60, TimeUnit.SECONDS)
-    }.build()!!
+    }.apply(httpClientBuilder).build()!!
+    val option = PenicillinOption.Builder().apply(optionBuilder).build()
     val pool = ScheduledThreadPoolExecutor(5)
 
     inline fun <reified T: JsonModel> getObject(url: String, authorizationType: AuthorizationType = AuthorizationType.OAuth1a, operation: PenicillinRequestBuilder.() -> Unit): ObjectAction<T> {
@@ -33,7 +34,7 @@ class Session(val consumerKey: String?, val consumerSecret: String?, val accessT
     inline fun getText(url: String, authorizationType: AuthorizationType = AuthorizationType.OAuth1a, operation: PenicillinRequestBuilder.() -> Unit): TextAction {
         return TextAction(PenicillinRequestBuilder(this, HTTPMethod.GET, url, authorizationType).apply(operation))
     }
-    inline fun <T: StreamHandler> getStream(url: String, authorizationType: AuthorizationType = AuthorizationType.OAuth1a, operation: PenicillinRequestBuilder.() -> Unit): StreamAction<T> {
+    inline fun <T: StreamListener> getStream(url: String, authorizationType: AuthorizationType = AuthorizationType.OAuth1a, operation: PenicillinRequestBuilder.() -> Unit): StreamAction<T> {
         return StreamAction(PenicillinRequestBuilder(this, HTTPMethod.GET, url, authorizationType).apply(operation))
     }
 
@@ -49,7 +50,7 @@ class Session(val consumerKey: String?, val consumerSecret: String?, val accessT
     inline fun postText(url: String, authorizationType: AuthorizationType = AuthorizationType.OAuth1a, operation: PenicillinRequestBuilder.() -> Unit): TextAction {
         return TextAction(PenicillinRequestBuilder(this, HTTPMethod.POST, url, authorizationType).apply(operation))
     }
-    inline fun <T: StreamHandler> postStream(url: String, authorizationType: AuthorizationType = AuthorizationType.OAuth1a, operation: PenicillinRequestBuilder.() -> Unit): StreamAction<T> {
+    inline fun <T: StreamListener> postStream(url: String, authorizationType: AuthorizationType = AuthorizationType.OAuth1a, operation: PenicillinRequestBuilder.() -> Unit): StreamAction<T> {
         return StreamAction(PenicillinRequestBuilder(this, HTTPMethod.POST, url, authorizationType).apply(operation))
     }
 
@@ -65,7 +66,7 @@ class Session(val consumerKey: String?, val consumerSecret: String?, val accessT
     inline fun deleteText(url: String, authorizationType: AuthorizationType = AuthorizationType.OAuth1a, operation: PenicillinRequestBuilder.() -> Unit): TextAction {
         return TextAction(PenicillinRequestBuilder(this, HTTPMethod.DELETE, url, authorizationType).apply(operation))
     }
-    inline fun <T: StreamHandler> deleteStream(url: String, authorizationType: AuthorizationType = AuthorizationType.OAuth1a, operation: PenicillinRequestBuilder.() -> Unit): StreamAction<T> {
+    inline fun <T: StreamListener> deleteStream(url: String, authorizationType: AuthorizationType = AuthorizationType.OAuth1a, operation: PenicillinRequestBuilder.() -> Unit): StreamAction<T> {
         return StreamAction(PenicillinRequestBuilder(this, HTTPMethod.DELETE, url, authorizationType).apply(operation))
     }
 }