+++ src/main/kotlin/jp/nephy/penicillin/model/DirectMessage.kt
@@ -2,7 +2,7 @@ package jp.nephy.penicillin.model
 
 import com.google.gson.JsonObject
 import jp.nephy.jsonkt.*
-import jp.nephy.penicillin.util.CreatedAt
+import jp.nephy.penicillin.model.special.CreatedAt
 
 
 class DirectMessage(override val json: JsonObject): JsonModel {