+++ src/main/kotlin/jp/nephy/penicillin/model/UserStreamListEvent.kt
@@ -1,0 +1,9 @@
+package jp.nephy.penicillin.model
+
+import com.google.gson.JsonObject
+import jp.nephy.jsonkt.byModel
+
+
+class UserStreamListEvent(json: JsonObject): UserStreamEvent(json) {
+    val targetObject by json.byModel<TwitterList>(key = "target_object")
+}