+++ src/main/kotlin/jp/nephy/penicillin/model/UserStreamStatusWithheld.kt
@@ -7,7 +7,7 @@ import jp.nephy.jsonkt.byString
 import jp.nephy.jsonkt.byStringList
 
 
-class StatusWithheld(override val json: JsonObject): JsonModel {
+class UserStreamStatusWithheld(override val json: JsonObject): JsonModel {
     val userId by json["status_withheld"].byLong("user_id")
     val id by json["status_withheld"].byLong
     val timestampMs by json["status_withheld"].byString("timestamp_ms")