+++ src/main/kotlin/jp/nephy/penicillin/streaming/StreamHandler.kt
@@ -1,7 +1,0 @@
-package jp.nephy.penicillin.streaming
-
-import com.google.gson.JsonObject
-
-interface StreamHandler {
-    fun handle(json: JsonObject)
-}